keysystem: rename keysystem/keycommon to keysystem/api
To make the fact that other components use this package clearer, rename
keysystem/keycommon to keysystem/api, and update uses as necessary.
