+++ building/src/keysystem/keyclient/actions/bootstrap/bootstrap_test.go
@@ -11,8 +11,8 @@ import (
 	"io/ioutil"
 	"keysystem/keyclient/config"
 	"keysystem/keyclient/state"
-	"keysystem/keycommon/reqtarget"
-	"keysystem/keycommon/server"
+	"keysystem/api/reqtarget"
+	"keysystem/api/server"
 	"keysystem/keyserver/authorities"
 	"net"
 	"net/http"
