+++ building/src/keysystem/keyclient/actions/keyreq/keyreq_test.go
@@ -14,8 +14,8 @@ import (
 	"io/ioutil"
 	"keysystem/keyclient/config"
 	"keysystem/keyclient/state"
-	"keysystem/keycommon/reqtarget"
-	"keysystem/keycommon/server"
+	"keysystem/api/reqtarget"
+	"keysystem/api/server"
 	"math/big"
 	"net"
 	"net/http"
