+++ building/src/keysystem/keyinitadmit/main/keyinitadmit.go
@@ -6,8 +6,8 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"encoding/pem"
-	"keysystem/keycommon/reqtarget"
-	"keysystem/keycommon/server"
+	"keysystem/api/reqtarget"
+	"keysystem/api/server"
 	"keysystem/keyserver/config"
 	"log"
 	"os"
