+++ building/src/keysystem/keyreq/main/keyreq.go
@@ -6,8 +6,8 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"io/ioutil"
-	"keysystem/keycommon/reqtarget"
-	"keysystem/keycommon/server"
+	"keysystem/api/reqtarget"
+	"keysystem/api/server"
 	"log"
 	"os"
 	"util/csrutil"
